cinema_environment/server_monolith/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-tables2==2.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
permission==0.4.1
Pillow==7.0.0
PyJWT==1.7.1
pymaging==0.1
pymaging-png==0.1
pytz==2019.3
qrcode==6.1
six==1.14.0
sqlparse==0.3.0
tablib==1.0.0
